+++ spec/ramda_spec.rb
@@ -58,6 +58,7 @@ describe Ramda do
   r(:join)
   r(:juxt)
   r(:keys)
+  # r(:keys_in)
   r(:last)
   r(:last_index_of)
   r(:length)